+++ packages/vue-docgen-api/src/script-handlers/mixinsHandler.ts
@@ -33,18 +33,20 @@ export default async function mixinsHandler(
 	// get all require / import statements
 	const mixinVarToFilePath = resolveRequired(astPath, mixinVariableNames)
 
-	resolveImmediatelyExportedRequire(pathResolver, mixinVarToFilePath)
+	await resolveImmediatelyExportedRequire(pathResolver, mixinVarToFilePath, opt.validExtends)
 
 	// get each doc for each mixin using parse
 	const files = new Map<string, string[]>()
 	for (const varName of Object.keys(mixinVarToFilePath)) {
 		const { filePath, exportName } = mixinVarToFilePath[varName]
-		const fullFilePath = pathResolver(filePath)
-		if (!/[\\/]node_modules[\\/]/.test(fullFilePath)) {
-			const vars = files.get(fullFilePath) || []
-			vars.push(exportName)
-			files.set(fullFilePath, vars)
-		}
+		filePath.forEach(p => {
+			const fullFilePath = pathResolver(p)
+			if (opt.validExtends(fullFilePath)) {
+				const vars = files.get(fullFilePath) || []
+				vars.push(exportName)
+				files.set(fullFilePath, vars)
+			}
+		})
 	}
 
 	await files.keys().reduce(async (_, fullFilePath) => {
@@ -66,12 +68,12 @@ export default async function mixinsHandler(
 					documentation
 				)
 				mixinVar.name = documentation.get('displayName')
+				documentation.set('displayName', null)
 			} catch (e) {
 				// eat the error
 			}
 		}
 	}, Promise.resolve())
-	documentation.set('displayName', null)
 }
 
 function getMixinsVariableNames(compDef: NodePath): string[] {