+++ packages/vue-docgen-api/src/utils/__tests__/adaptExportsToIEV.ts
@@ -7,13 +7,22 @@ describe('adaptRequireWithIEV', () => {
 	let set: ImportedVariableSet
 	let mockResolver: jest.Mock
 	beforeEach(() => {
-		set = { test: { filePath: 'my/path', exportName: 'exportIt' } }
+		set = { test: { filePath: ['my/path'], exportName: 'exportIt' } }
 		mockResolver = jest.fn()
 	})
 
-	it('should call the resolver', () => {
-		adaptRequireWithIEV(mockResolver, set)
+	it('should call the resolver', async done => {
+		await adaptRequireWithIEV(mockResolver, set, () => true)
 
 		expect(mockResolver).toHaveBeenCalledWith('my/path')
+		done()
+	})
+
+	it('should not resolve anything if multiple path in filePath', async done => {
+		set.test.filePath.push('baz')
+		await adaptRequireWithIEV(mockResolver, set, () => true)
+
+		expect(mockResolver).not.toHaveBeenCalledWith()
+		done()
 	})
 })