+++ packages/vue-docgen-api/tests/components/extended/InputTextDocumented.vue
@@ -1,17 +1,20 @@
 <template>
-  <div>
-    <!--How to include the question here???-->
-    <input :placeholder="placeholder">
-  </div>
+	<div>
+		<!--How to include the question here???-->
+		<input :placeholder="placeholder" />
+	</div>
 </template>
 <script>
 import Base from './Base.vue'
+import { multi, other } from '../../mixins'
+import { deep } from '../../mixins'
 
 /**
  * Description InputTextDocumented
  */
 export default {
 	extends: Base,
+	mixins: [multi, other, deep],
 	props: {
 		placeholder: {
 			type: String