+++ packages/vue-docgen-api/tests/mixins/multiNamed.js
@@ -1,0 +1,7 @@
+export const multi = {
+	props: {
+		importedProp: String
+	}
+}
+
+export { default as deep } from './another/deep'